feat: redirect from form register
