+++ e-commerce/src/assets/sass/base/_body.scss
@@ -82,7 +82,9 @@ button:disabled {
   background-color: $Grey-dark-pagination;
   cursor: none;
 }
-
+.none {
+  display: none;
+}
 @media (max-width: 700px) {
   //   body {
   //   @include font($size: 1.6rem, $height: 150%, $weight: 400);
